+++ src/features/label/ComponentWithLabel/ComponentWithLabel.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import type { ReactNode } from 'react';
 
-import classes from 'src/layout/ComponentWithLabel.module.css';
-import { LabelContent } from 'src/layout/LabelContent';
+import classes from 'src/features/label/ComponentWithLabel/ComponentWithLabel.module.css';
+import { LabelContent } from 'src/features/label/LabelContext/LabelContent';
 import type { ILabelSettings } from 'src/layout/common.generated';
 
 export interface ComponentWithLabelProps {
